docs: revise QURI Parts integration tutorial per review
Rename quri_parts_executor.{py,ipynb} -> quri_parts_support.{py,ipynb}
(en/ja) so the file name matches the H1 "QURI Parts Support" and
reflects that the article covers both transpile and execute. Update
docs/{en,ja}/integration/index.md links and tests/docs/test_tutorials.py
(TUTORIAL_PATTERNS paths, OPTIONAL_SKIP_MODULES key).

Restructure the tutorial:
- Write the QAOA ansatz directly as small qkernel-decorated building
  blocks (superposition / cost_layer / mixer_layer / qaoa_ansatz)
  instead of going through QAOAConverter, so the Qamomile-side IR
  stays inspectable via kernel.draw(). Build the cost Hamiltonian
  with qamomile.observable directly and decode samples via
  BinaryModel.decode_from_sampleresult; the model object is used
  purely as a problem container.
- Add QURI Parts' native draw_circuit() after get_first_circuit() so
  readers can visually confirm the returned object really is a
  LinearMappedParametricQuantumCircuit.
- Replace the code-block-only "Customizing the sampler" section with
  a working NoiseSimulator example (DepolarizingNoise + Qulacs
  noisesimulator sampler) that compares noiseless vs noisy
  sample-mean energies at the same optimized parameters, so the
  effect of the custom sampler is visible.

Notebooks are re-synced from the .py and executed end-to-end with
QAMOMILE_DOCS_TEST unset (full settings: shots=2000, maxiter=100).
