+++ docs/en/integration/index.md
@@ -9,4 +9,4 @@ Notes on using Qamomile together with external libraries and quantum platforms.
 ## All articles
 
 - [qBraid Executor](qbraid_executor) — Run Qiskit circuits on qBraid-supported devices
-- [QURI Parts Executor](quri_parts_executor) — Transpile to QURI Parts and run on a Qulacs state-vector simulator
+- [QURI Parts Support](quri_parts_support) — Transpile to QURI Parts and run on a Qulacs state-vector simulator
